removes background tasks
not needed since config and fake request are done in background handleer of SDK
